+++ spec/threatinator/event_builder_spec.rb
@@ -17,7 +17,7 @@ describe Threatinator::EventBuilder do
       event_builder.reset
       expect {
         event_builder.build
-      }.to raise_error(Threatinator::Exceptions::InvalidAttributeError)
+      }.to raise_error(Threatinator::Exceptions::EventBuildError)
     end
 
     it "resets the fqdns" do
@@ -51,6 +51,7 @@ describe Threatinator::EventBuilder do
       expect(event1.type).to eq(:c2)
     end
   end
+
   describe "#add_ipv4(ipv4)" do
     it "adds the provided ipv4s to built events" do
       event_builder.add_ipv4('1.2.3.4')
@@ -59,12 +60,25 @@ describe Threatinator::EventBuilder do
       expect(event1.ipv4s).to contain_exactly('1.2.3.4', '8.8.8.8')
     end
   end
+
   describe "#add_fqdn(fqdn)" do
     it "adds the provided fqdns to built events" do
       event_builder.add_ipv4('google.com')
       event_builder.add_ipv4('yahoo.com')
       event1 = event_builder.build
       expect(event1.ipv4s).to contain_exactly('google.com', 'yahoo.com')
+    end
+  end
+
+  describe "#add_url(url)" do
+    it "converts the provided URLs strings into Addressable::URI objects and adds them to the built events" do
+      event_builder.add_url('http://google.com/foo/bar')
+      event_builder.add_url('http://yahoo.com')
+      event = event_builder.build
+      expect(event.urls).to contain_exactly(
+        ::Addressable::URI.parse('http://google.com/foo/bar'),
+        ::Addressable::URI.parse('http://yahoo.com')
+      )
     end
   end
 
@@ -84,6 +98,25 @@ describe Threatinator::EventBuilder do
       expect(event1).to be == event2
     end
 
+    context "when an added URL is not parseable as a URI" do
+      it "raises EventBuildError" do
+        event_builder.type = :c2
+        event_builder.add_url(1234)
+        expect {
+          event_builder.build
+        }.to raise_error(Threatinator::Exceptions::EventBuildError)
+      end
+    end
+
+    context "when an added URL is not absolute" do
+      it "raises EventBuildError" do
+        event_builder.type = :c2
+        event_builder.add_url("/foo/bar")
+        expect {
+          event_builder.build
+        }.to raise_error(Threatinator::Exceptions::EventBuildError)
+      end
+    end
 
   end
 end